requirements.txt
requests
python-dotenv
langchain
langgraph
fastapi
uvicorn[standard]
pydantic
gradio
httpx
